# docs/swagger.yaml
definitions:
...
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - name
    - role
    type: object
  auth.ResendRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
info:
  contact: {}
paths:
...
  /v1/login:
    post:
      description: Login
      parameters:
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login
      tags:
      - Auth
  /v1/logout:
    post:
      description: Logout
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Logout
      tags:
      - Auth
...
  /v1/register:
    post:
      description: Register
      parameters:
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register [SUPER ADMIN ONLY]
      tags:
      - Auth
  /v1/resend:
    post:
      description: Resend email for HR verification
      parameters:
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/auth.ResendRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Resend Email
      tags:
      - Auth
...
  /v1/verify:
    get:
      description: Verify Email
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      - description: Token
        in: query
        name: token
        required: true
        type: string
      responses:
        "302":
          description: Found
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Verify Email
      tags:
      - Auth
swagger: "2.0"