# src/entity/v1/http/auth/response.go
package auth

type UserResponse struct {
	ID    string `json:"id"`
	Name  string `json:"name"`
	Email string `json:"email"`
	Role  string `json:"role"`
}

type LoginResponse struct {
	UserResponse
	AccessToken string `json:"access_token"`
}